pyproject.toml
[tool.poetry]
name = "ronpari"
version = "0.1.0"
description = ""
authors = ["Artem Basenko"]

[tool.poetry.dependencies]
python = "^3.10"
"MangaDex.py" = "^2.0.10"
click = "^8.1.3"
rich = "^12.5.1"
rich-click = "^1.5.1"
tinydb = "^4.7.0"
mangadex = "^2.5.2"
opencv-python = "^4.6.0"
imutils = "^0.5.4"
screeninfo = "^0.8"
Pillow = "^9.2.0"
pretty-errors = "^1.2.25"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
poethepoet = "^0.15.0"
bpython = "^0.22.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ronpari = "ronpari.cli:manga_cli"

[tool.poe.tasks]
test = "pytest -v"

